pyproject.toml
[project]
name = "asap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
license = "Apache-2.0"
authors = [
    {name = "Ethan Katz"}
]
requires-python = ">=3.11"
dependencies = [
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "zaber-motion>=8.4.0",
]

[dependency-groups]
dev = [
    "basedpyright>=1.38.3",
    "docsig>=0.81.1",
    "pytest>=9.0.2",
    "ruff>=0.15.7",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.basedpyright]
include = ["src", "tests"]
exclude = [".venv", "__pycache__"]
typeCheckingMode = "recommended"

[tool.ruff]
lint.select = ["ALL"]
